requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.2
django-rest-swagger==2.2.0
djangorestframework==3.12.1
idna==2.10
isort==5.5.4
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
Pillow==7.2.0
pycodestyle==2.6.0
pylint==2.6.0
pytz==2020.1
requests==2.24.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.10
wrapt==1.12.1
